OpenXR and input fixes, updated ImGui
- More generic OpenXR action bindings
- Fixed use of jolt shape settings
- Updated ImGui version
- Implemented HUD grabbing
- Simplified flow of input events
- Improved commands
- Moved safe_normalize_cross() into erhe toolkit math_util
- Most uses of double were changed to use float
- erhe::application::Command_context -> Input_arguments
- Viewport_window::to_scene_content() -> viewport_from_window()
- Controller upgraded from a ray to mat4 transform
- Added Node::set_node_from_world()
- Added ERHE_HASH() for compile time string hash

+++ src/editor/rendertarget_imgui_viewport.hpp
@@ -81,8 +81,8 @@ public:
 private:
     Rendertarget_mesh* m_rendertarget_mesh{nullptr};
     glm::vec4          m_clear_color {0.0f, 0.0f, 0.0f, 0.2f};
-    double             m_last_mouse_x{0.0};
-    double             m_last_mouse_y{0.0};
+    float              m_last_mouse_x{0.0f};
+    float              m_last_mouse_y{0.0f};
 };
 
 } // namespace erhe::application
