+++ src/editor/tools/grid.hpp
@@ -33,7 +33,7 @@ static constexpr const char* grid_plane_type_strings[] =
     "Node"
 };
 
-auto get_plane_transform(Grid_plane_type plane_type) -> glm::dmat4;
+auto get_plane_transform(Grid_plane_type plane_type) -> glm::mat4;
 
 class Grid
     : public erhe::scene::Node_attachment
@@ -47,14 +47,14 @@ public:
 
     // Public API
     [[nodiscard]] auto get_name           () const -> const std::string&;
-    [[nodiscard]] auto snap_world_position(const glm::dvec3& position_in_world) const -> glm::dvec3;
-    [[nodiscard]] auto snap_grid_position (const glm::dvec3& position_in_grid ) const -> glm::dvec3;
-    [[nodiscard]] auto world_from_grid    () const -> glm::dmat4;
-    [[nodiscard]] auto grid_from_world    () const -> glm::dmat4;
+    [[nodiscard]] auto snap_world_position(const glm::vec3& position_in_world) const -> glm::vec3;
+    [[nodiscard]] auto snap_grid_position (const glm::vec3& position_in_grid ) const -> glm::vec3;
+    [[nodiscard]] auto world_from_grid    () const -> glm::mat4;
+    [[nodiscard]] auto grid_from_world    () const -> glm::mat4;
     [[nodiscard]] auto intersect_ray(
-        const glm::dvec3& ray_origin_in_world,
-        const glm::dvec3& ray_direction_in_world
-    ) -> std::optional<glm::dvec3>;
+        const glm::vec3& ray_origin_in_world,
+        const glm::vec3& ray_direction_in_world
+    ) -> std::optional<glm::vec3>;
 
     void render(const Render_context& context);
     void imgui ();
@@ -62,8 +62,8 @@ public:
 private:
     std::string     m_name;
     Grid_plane_type m_plane_type      {Grid_plane_type::XZ};
-    double          m_rotation        {0.0}; // Used only if plane type != node
-    glm::dvec3      m_center          {0.0}; // Used only if plane type != node
+    float           m_rotation        {0.0f}; // Used only if plane type != node
+    glm::vec3       m_center          {0.0f}; // Used only if plane type != node
     bool            m_enable          {true};
     bool            m_see_hidden_major{false};
     bool            m_see_hidden_minor{false};
@@ -74,8 +74,8 @@ private:
     float           m_minor_width     {2.0f};
     glm::vec4       m_major_color     {0.0f, 0.0f, 0.0f, 0.729f};
     glm::vec4       m_minor_color     {0.0f, 0.0f, 0.0f, 0.737f};
-    glm::dmat4      m_world_from_grid {1.0f};
-    glm::dmat4      m_grid_from_world {1.0f};
+    glm::mat4       m_world_from_grid {1.0f};
+    glm::mat4       m_grid_from_world {1.0f};
 };
